backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
PyJWT==2.8.0
pytz==2023.3
qrcode==7.4.2
Pillow==10.0.1
gunicorn==21.2.0
python-dotenv==1.0.0